+++ packages/media/src/lib/image/BaseImagePlugin.ts
@@ -2,6 +2,7 @@ import {
   type PluginConfig,
   bindFirst,
   createTSlatePlugin,
+  isSlatePluginElement,
 } from '@udecode/plate';
 
 import type { MediaPluginOptions, TMediaElement } from '../media';
@@ -55,10 +56,16 @@ export const BaseImagePlugin = createTSlatePlugin<ImageConfig>({
     parsers: {
       html: {
         deserializer: {
-          parse: ({ element }) => ({
-            type: plugin.node.type,
-            url: element.getAttribute('src'),
-          }),
+          parse: ({ element, type }) => {
+            if (isSlatePluginElement(element as HTMLElement, type)) {
+              return;
+            }
+
+            return {
+              type: plugin.node.type,
+              url: element.getAttribute('src'),
+            };
+          },
           rules: [
             {
               validNodeName: 'IMG',