+++ templates/plate-playground-template/src/components/editor/plugins/block-selection-plugins.ts
@@ -3,15 +3,17 @@
 import { BlockSelectionPlugin } from '@udecode/plate-selection/react';
 
 export const blockSelectionPlugins = [
-  BlockSelectionPlugin.configure({
-    inject: {
-      excludeBelowPlugins: ['tr'],
-      excludePlugins: ['table', 'code_line', 'column_group', 'column'],
-    },
+  BlockSelectionPlugin.configure(({ editor }) => ({
     options: {
       enableContextMenu: true,
+      isSelectable: (element, path) => {
+        return (
+          !['code_line', 'column', 'td'].includes(element.type) &&
+          !editor.api.block({ above: true, at: path, match: { type: 'tr' } })
+        );
+      },
     },
-  }),
+  })),
 ] as const;
 
 export const blockSelectionReadOnlyPlugin = BlockSelectionPlugin.configure({